:recycle: (pc-15004) move useProfileOptions to api. delete duplicate tests
