--- src/features/identityCheck/utils/__mocks__/useProfileOptions.tsx
@@ -1,8 +1,0 @@
-import { SchoolTypesSnap } from 'features/identityCheck/pages/profile/fixtures/mockedSchoolTypes'
-
-export const useProfileOptions = jest.fn(() => {
-  return {
-    schoolTypes: SchoolTypesSnap.school_types,
-    activities: SchoolTypesSnap.activities,
-  }
-})
